Feat: Updated version to new model
