+++ src/datamarts/collections/regulon_datamart.py
@@ -1,7 +1,7 @@
 from src.datamarts.domain.general.biological_base import BiologicalBase
 import multigenomic_api
 
-from src.datamarts.domain.regulon_datamart.transcription_factor import TranscriptionFactor
+from src.datamarts.domain.regulon_datamart.regulator import Regulator
 from src.datamarts.domain.regulon_datamart.regulates import Regulates
 from src.datamarts.domain.regulon_datamart.terms import Terms
 from src.datamarts.domain.regulon_datamart.regulatory_interactions import RegulatoryInteractions
@@ -14,43 +14,58 @@ class RegulonDatamarts:
 
     @property
     def objects(self):
-        regulator_objects = multigenomic_api.transcription_factors.get_all()
-        for regulator_object in regulator_objects:
-            print(regulator_object.id)
-            regulon_datamart = RegulonDatamarts.RegulonDatamart(regulator_object)
+        regulator_objects = get_all_regulators(multigenomic_api.regulatory_interactions.get_all())
+        tf_objects = multigenomic_api.transcription_factors.get_all()
+        """for tf_obj in tf_objects:
+            print(tf_obj.id)
+            tf_obj["regulator_type"] = "transcriptionFactor"
+            regulon_datamart = RegulonDatamarts.RegulonDatamart(tf_obj)
+            yield regulon_datamart"""
+        for reg_obj in regulator_objects[0:1]:
+            print(reg_obj.id)
+            regulator = reg_obj
+            if reg_obj.type == "srna":
+                regulator = multigenomic_api.products.find_by_id(reg_obj.id)
+            if reg_obj.type == "compound":
+                regulator = multigenomic_api.regulatory_continuants.find_by_id(reg_obj.id)
+            regulator["regulator_type"] = reg_obj.type
+            regulon_datamart = RegulonDatamarts.RegulonDatamart(regulator)
             yield regulon_datamart
         del regulator_objects
 
     class RegulonDatamart:
 
-        def __init__(self, transcription_factor):
-            self.id = transcription_factor.id
-            self.transcription_factor = transcription_factor
-            self.terms = transcription_factor
-            self.regulates = transcription_factor
-            self.regulatory_interactions = transcription_factor.active_conformations
-            self.alignmentMatrix = transcription_factor.id
-            self.evolutionaryConservation = transcription_factor.id
-            self.organism = transcription_factor.organisms_id
-            self.summary = [self.regulates, self._regulatory_interactions]
+        def __init__(self, regulator):
+            self.id = regulator.id
+            self.regulator = regulator
+            self.terms = regulator
+            # self.regulates = regulator
+            self.regulatory_interactions = regulator
+            # self.alignmentMatrix = regulator.id
+            # self.evolutionaryConservation = regulator.id
+            # self.organism = regulator.organisms_id
+            # self.summary = [self.regulates, self._regulatory_interactions]
 
         @property
-        def transcription_factor(self):
-            return self._transcription_factor
+        def regulator(self):
+            return self._regulator
 
-        @transcription_factor.setter
-        def transcription_factor(self, transcription_factor):
-            transcription_factor = TranscriptionFactor(transcription_factor)
-            self._transcription_factor = transcription_factor.to_dict()
+        @regulator.setter
+        def regulator(self, regulator):
+            regulator = Regulator(regulator)
+            self._regulator = regulator.to_dict()
 
         @property
         def terms(self):
             return self._terms
 
         @terms.setter
-        def terms(self, transcription_factor):
-            terms = Terms(transcription_factor)
-            self._terms = terms.to_dict()
+        def terms(self, regulator):
+            self._terms = []
+            terms = []
+            if regulator.regulator_type == "transcriptionFactor":
+                terms = Terms(regulator).to_dict()
+            self._terms = terms
 
         @property
         def regulates(self):
@@ -66,17 +81,16 @@ class RegulonDatamarts:
             return self._regulatory_interactions
 
         @regulatory_interactions.setter
-        def regulatory_interactions(self, tf_active_conformations):
+        def regulatory_interactions(self, regulator):
             self._regulatory_interactions = []
-            for active_conformation in tf_active_conformations:
-                regulatory_interactions = multigenomic_api.regulatory_interactions.find_by_regulator_id(
-                    active_conformation.id)
+            if regulator.regulator_type == "transcriptionFactor":
+                for active_conformation in regulator.active_conformations:
+                    regulatory_interactions = multigenomic_api.regulatory_interactions.find_by_regulator_id(
+                        active_conformation.id)
+                    self._regulatory_interactions = get_ri_objects(regulatory_interactions, self._regulatory_interactions)
+            else:
+                regulatory_interactions = multigenomic_api.regulatory_interactions.find_by_regulator_id(regulator.id)
                 self._regulatory_interactions = get_ri_objects(regulatory_interactions, self._regulatory_interactions)
-                if active_conformation.type == "regulatoryComplex":
-                    reg_complex = multigenomic_api.regulatory_complexes.find_by_id(active_conformation.id)
-                    for reg_cont_id in reg_complex.regulatory_continuants_ids:
-                        regulatory_interactions = multigenomic_api.regulatory_interactions.find_by_regulator_id(reg_cont_id)
-                        self._regulatory_interactions = get_ri_objects(regulatory_interactions, self._regulatory_interactions)
 
         @property
         def organism(self):
@@ -106,15 +120,15 @@ class RegulonDatamarts:
         def to_dict(self):
             regulon_datamart = {
                 "_id": self.id,
-                "transcriptionFactor": self.transcription_factor,
-                "terms": self.terms,
-                "regulates": self.regulates,
+                "regulator": self.regulator,
+                # "terms": self.terms,
+                # "regulates": self.regulates,
                 "regulatoryInteractions": self.regulatory_interactions,
                 # "alignmentMatrix": [],
                 # "evolutionaryConservation":[],
-                "organism": self.organism,
-                "summary": self.summary,
-                "allCitations": BiologicalBase.get_all_citations()
+                # "organism": self.organism,
+                # "summary": self.summary,
+                # "allCitations": BiologicalBase.get_all_citations()
             }
             return regulon_datamart
 
@@ -134,3 +148,21 @@ def get_ri_objects(reg_ints, ri_list):
         if reg_int not in ri_list:
             ri_list.append(reg_int)
     return ri_list
+
+
+def get_all_regulators(reg_ints):
+    regulators = []
+    for ri in reg_ints:
+        if ri.regulator.type == "regulatoryContinuant":
+            regulator = ri.regulator
+            regulator["type"] = "compound"
+            if ri.regulator not in regulators:
+                regulators.append(ri.regulator)
+        if ri.regulator.type == "product":
+            product = multigenomic_api.products.find_by_id(ri.regulator.id)
+            if product.type == "small RNA":
+                regulator = ri.regulator
+                regulator["type"] = "srna"
+                if ri.regulator not in regulators:
+                    regulators.append(ri.regulator)
+    return regulators
